Re-organize code for ease of use
Added a "SquiggleDraw" directory with the sketch and a "data" folder.
This prevents Processing from requesting that you move the sketch into
a like-named folder. (And, thus, prevents the situation where the
Carson image cannot be found because it wasn't moved along with the
sketch.)

Added note to the readme about ControlP5 dependency.

Both of these changes are made in response to multiple instances
reported to us of people running into these issues.