Issues #14, #15
Rework how row reversing is done.

Rows are now calculated identically if reversed or not. We build arrays of X and Y points for each row. Then, we reverse those lists -- to only change the order that the vertices are added to the curve -- if the row should be drawn in reverse order.